Get from Amazon S3
This uses an http request to get the latest greenlit hash for a
specific version, and

Rather than use a Node fetch() wrapper or similar, this uses the
GitHub-Actions-speific http module, to follow the convention of what
it is using for downloads in general.  It's not clear what the benefit
of that is, but it means that if anything github-actions-specific
related to http requests of the outside world comes up...they will
be fixing it there, and this will get the fix.