Improve parameterization of Python script
The timeout was not being actually passed to the Python helper
script because the CLI parameter processing was clunky.  This
goes ahead and switches to use the Python argument parser, to
handle the timeout as well as a parameter for taking a screenshot.

As a convenience, the screenshot is implicitly uploaded as an
artifact so you don't have to have a separate action for that.

This updates the method of noticing success to be based on the
setInterval function instead of setTimeout, so it polls the
browser to see if the exit condition has been reached.  It also
catches the exception caused by the Marionette timeout in order
to still be able to take a screenshot and shut down the browser.